敵の基底クラスを作成　基底クラスを派生させて通常ファンを作成 #79
