+++ Assets/Prefabs/Enemy/Enemy.prefab
@@ -9,7 +9,7 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 5968483811882916910}
-  - component: {fileID: 1060883709}
+  - component: {fileID: 7207804476372920346}
   - component: {fileID: -3143044023716922937}
   - component: {fileID: 5968483811882916911}
   - component: {fileID: 336802384846892835}
@@ -36,7 +36,7 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &1060883709
+--- !u!114 &7207804476372920346
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -45,22 +45,22 @@ MonoBehaviour:
   m_GameObject: {fileID: 5968483811882916908}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 1e230a272d83991498961b02d2f21dac, type: 3}
+  m_Script: {fileID: 11500000, guid: 55f1c565b5844174093665ed4a401d4a, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   _enemySpped: 5
-  _fadedSpeed: 1.5
+  _fadedSpeed: 1
   _resultTimes:
   - 6
-  - 4
-  - 2
-  - 0.5
+  - 5
+  - 3
+  - 1
   - 0
-  _addScoreValue: 1
-  _deadMoveTime: 2
-  _minScale: 0.1
+  _requestUIArray:
+  - {fileID: 7277948641732065839}
   _explosionEffect: {fileID: 5047219227802861214, guid: 5315674b013e09d4088b189a94135c4c, type: 3}
-  _requestUIController: {fileID: 7277948641732065839}
+  _deadMoveTime: 1
+  _minScale: 0.3
   _trajectoryParent:
   - {fileID: 5039499519867245985, guid: 2d33b4343f350624e8d45b8342a5ecc0, type: 3}
   - {fileID: 5039499519867245985, guid: bc1a3faa40214fa4f9ea5edd3e343288, type: 3}
