+++ Assets/Scripts/Enemy/EnemyBase.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 1e230a272d83991498961b02d2f21dac
+guid: 801c385d3f130cc41b86bd59a77a9f74
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
